demo/requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
codecarbon
